requirements.txt
boto3==1.9.243
botocore==1.12.243
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-extra-fields==2.0.2
django-filter==2.2.0
django-rest-auth==0.9.5
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
drf-extensions==0.5.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
Pillow==6.2.0
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
sendgrid==6.1.0
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
urllib3==1.25.6
WeasyPrint==51
webencodings==0.5.1
whitenoise==4.1.4
